feat: implement comprehensive staff module migrations, authorization policies, and core services for school management system

+++ Modules/Staff/Database/Migrations/2026_04_01_000006_create_staff_leave_details_table.php
@@ -10,8 +10,8 @@ return new class extends Migration
     {
         Schema::create('staff_leave_details', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('staff_id');
-            $table->integer('leave_type_id');
+            $table->unsignedBigInteger('staff_id');
+            $table->unsignedBigInteger('leave_type_id');
             $table->string('alloted_leave', 100);
         });
     }
