+++ Modules/Staff/Database/Migrations/2026_04_01_000012_create_staff_timeline_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('staff_timeline', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('staff_id');
+            $table->unsignedBigInteger('staff_id');
             $table->string('title', 200);
             $table->date('timeline_date');
             $table->string('description', 300);
